services/service-mesh/mesh_config.py
#!/usr/bin/env python3
"""
Service Mesh Configuration Loader
"""
import yaml
import logging
import time
import os

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def load_config():
    """Load the mesh configuration from YAML file"""
    try:
        with open('mesh_config.yaml', 'r') as file:
            config = yaml.safe_load(file)
            logger.info(f"Loaded mesh configuration: {config}")
            return config
    except Exception as e:
        logger.error(f"Error loading mesh configuration: {str(e)}")
        return None

def main():
    """Main function to run the service mesh configuration"""
    logger.info("Starting service mesh configuration")
    
    # Load the configuration
    config = load_config()
    if not config:
        logger.error("Failed to load configuration. Exiting.")
        return
    
    # In a real implementation, this would apply the configuration to the service mesh
    logger.info("Service mesh configuration applied successfully")
    
    # Keep the container running
    while True:
        logger.info("Service mesh running...")
        time.sleep(60)

if __name__ == "__main__":
    main()