services/service-mesh/requirements.txt
# Service Mesh Dependencies
fastapi==0.103.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.1
PyYAML==6.0
loguru==0.7.0
prometheus-client==0.16.0
redis==4.5.5
uuid==1.30